requirements.txt
python-dotenv==0.14.0
tweepy==3.10.0
beautifulsoup4==4.9.3
pandas==1.2.0
Pillow==7.0.0
regex==2020.11.13
requests==2.22.0
pathlib==1.0.1
